requirements.txt
asgiref==3.8.1
brotlicffi==1.1.0.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
Django==4.2.16
django-cors-headers==4.6.0
djangorestframework==3.15.2
drf-yasg==1.21.8
Faker==25.9.2
idna==3.10
inflection==0.5.1
packaging==24.1
pycparser==2.22
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
